Refactor version bump code.
