+++ alvr/common/src/data/version.rs
@@ -14,38 +14,3 @@ pub fn is_version_compatible(version: &str, requirement: &str) -> StrResult<bool
     let requirement = trace_err!(VersionReq::parse(requirement))?;
     Ok(requirement.matches(&version))
 }
-
-pub fn bumped_versions(server_version: Option<String>, client_version: Option<String>) -> Result<(String, String), String>{
-    let server_version = server_version.as_deref().unwrap_or(ALVR_SERVER_VERSION);
-    let client_version = client_version.as_deref().unwrap_or(ALVR_CLIENT_VERSION);
-
-    let server_req: String = format!(">={}", ALVR_SERVER_VERSION).into();
-    let server_req = trace_err!(VersionReq::parse(&server_req))?;
-    let server_version = trace_err!(Version::parse(server_version))?;
-
-    let client_req: String = format!(">={}", ALVR_CLIENT_VERSION).into();
-    let client_req = trace_err!(VersionReq::parse(&client_req))?;
-    let client_version = trace_err!(Version::parse(client_version))?;
-
-    if client_version.major != server_version.major {
-        return Err("Bumped versions need to have the same major version!".to_owned());
-    }
-
-    let server_version = if server_req.matches(&server_version) {
-        format!("{}", server_version)
-    } else {
-        return Err(format!("Cannot bump server version: {} -> {}", ALVR_SERVER_VERSION, server_version));
-    };
-
-    let client_version = if client_req.matches(&client_version) {
-        format!("{}", client_version)
-    } else {
-        return Err(format!("Cannot bump client version: {} -> {}", ALVR_CLIENT_VERSION, client_version));
-    };
-
-    if client_version == ALVR_CLIENT_VERSION && server_version == ALVR_SERVER_VERSION {
-        Err("Didn't bump any version!".to_owned())
-    } else {
-        Ok((client_version, server_version))
-    }
-}
